server/requirements.txt
pandas
numpy
scikit-learn
yfinance
joblib
pandas_ta
fastapi
uvicorn
pydantic
lxml
python-dotenv
google-generativeai
newspaper3k
lxml_html_clean